Add destructive migration to database downgrade
This might not really be scalable as we need to add one more drop table in `onDowngrade` we have new table. A better solution might be to query the sqlite master table for tables names then recursively drop them one by one. Since we only have 3 tables for now, I made a conscious decision not to do early abstraction.

Fixes #216